Support 32 match types in UidOwnerValue rule
Match type is also used in configuration_map at index
UID_RULES_CONFIGURATION_KEY.
However, this commit does not extend configuration_map and we can
not still use more than 8 match type in configuration_map.

Bug: 206482423
Test: m
(cherry picked from commit 42edc60627899989335dafa568f6bf724ea1a104)
Merged-In: I0f20cc8034551806b5cb3da322a0ea7861983095

+++ bpf_progs/netd.c
@@ -194,7 +194,7 @@ static inline int bpf_owner_match(struct __sk_buff* skb, uint32_t uid, int direc
     BpfConfig enabledRules = getConfig(UID_RULES_CONFIGURATION_KEY);
 
     UidOwnerValue* uidEntry = bpf_uid_owner_map_lookup_elem(&uid);
-    uint8_t uidRules = uidEntry ? uidEntry->rule : 0;
+    uint32_t uidRules = uidEntry ? uidEntry->rule : 0;
     uint32_t allowed_iif = uidEntry ? uidEntry->iif : 0;
 
     if (enabledRules) {
